feat: add files for storing heap
DSA CT heap section add files for storing our heap. this is really cool actually - so we dont save it if its currently empty, but we do always load it on each program startup. not sure if i would want to do that though, but i think we also need a button for clearing our heap now since user might not wanna load his previous session