feat: dont wanna be doing it now
but in a nutshell, spotify doesnt provide lyrics for songs, which i should've understood by this point cause there are a ton of songs without lyrics. and the genius thing requires an API key. ill maybe look into it and if its not paid and not limited in any way, ill use that, but for now ill just chill. the policy is completed and i dont wanna be doing anything beyond that. have a great day and stay Blessed. Glory to Jesus Christ 🙏